+++ mobile/android/android-components/components/concept/engine/src/test/java/mozilla/components/concept/engine/EngineTest.kt
@@ -47,6 +47,17 @@ class EngineTest {
     }
 
     @Test
+    fun `invokes default functions on trackingProtectionExceptionStore`() {
+        var wasExecuted = false
+        try {
+            testEngine.trackingProtectionExceptionStore
+        } catch (_: Exception) {
+            wasExecuted = true
+        }
+        assertTrue(wasExecuted)
+    }
+
+    @Test
     fun `invokes error callback if webextensions not supported`() {
         var exception: Throwable? = null
         testEngine.installWebExtension("my-ext", "resource://path", onError = { _, e -> exception = e })