+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/TrackingProtectionUseCasesTest.kt
@@ -8,6 +8,7 @@ import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.content.blocking.TrackerLog
+import mozilla.components.concept.engine.content.blocking.TrackingProtectionExceptionStorage
 import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.whenever
@@ -17,6 +18,7 @@ import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
+import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 
 class TrackingProtectionUseCasesTest {
@@ -24,11 +26,13 @@ class TrackingProtectionUseCasesTest {
     private val mockSessionManager = mock<SessionManager>()
     private val mockSession = mock<EngineSession>()
     private val mockEngine = mock<Engine>()
+    private val mockStore = mock<TrackingProtectionExceptionStorage>()
     private val useCases = TrackingProtectionUseCases(mockSessionManager, mockEngine)
 
     @Before
     fun setup() {
         whenever(mockSessionManager.getEngineSession(any())).thenReturn(mockSession)
+        whenever(mockEngine.trackingProtectionExceptionStore).thenReturn(mockStore)
     }
 
     @Test
@@ -86,5 +90,66 @@ class TrackingProtectionUseCasesTest {
         assertNull(trackersLog)
         assertTrue(onErrorCalled)
         assertFalse(onSuccessCalled)
+    }
+
+    @Test
+    fun `add exception`() {
+        useCases.addException(mock())
+        verify(mockSessionManager).getEngineSession(any())
+        verify(mockStore).add(any())
+    }
+
+    @Test
+    fun `add exception with a null engine session will not call the store`() {
+        whenever(mockSessionManager.getEngineSession(any())).thenReturn(null)
+
+        useCases.addException(mock())
+        verify(mockStore, times(0)).add(any())
+    }
+
+    @Test
+    fun `remove exception`() {
+        useCases.removeException(mock())
+        verify(mockSessionManager).getEngineSession(any())
+        verify(mockStore).remove(any())
+    }
+
+    @Test
+    fun `remove exception with a null engine session will not call the store`() {
+        whenever(mockSessionManager.getEngineSession(any())).thenReturn(null)
+
+        useCases.removeException(mock())
+        verify(mockStore, times(0)).remove(any())
+    }
+
+    @Test
+    fun `removeAll exceptions`() {
+        useCases.removeAllExceptions()
+        verify(mockStore).removeAll()
+    }
+
+    @Test
+    fun `contains exception`() {
+        useCases.containsException(mock()) {}
+        verify(mockStore).contains(any(), any())
+    }
+
+    @Test
+    fun `contains exception with a null engine session will not call the store`() {
+        var contains = true
+        whenever(mockSessionManager.getEngineSession(any())).thenReturn(null)
+
+        useCases.containsException(mock()) {
+            contains = it
+        }
+
+        assertFalse(contains)
+        verify(mockStore, times(0)).contains(any(), any())
+    }
+
+    @Test
+    fun `fetch exceptions`() {
+        useCases.fetchExceptions {}
+        verify(mockStore).fetchAll(any())
     }
 }